nominating exposure functions moved to src/renderer/callbacks/nominating.ts
